// jungsi-frontend/src/stores/server/features/explore/early-comprehensive/interfaces.ts
export interface IExploreEarlyComprehensiveDetailResponse
  extends IRecruitmentUnit {
  admission: {
    id: number;
    name: string;
    year: number;
    basic_type: "일반" | "특별";
    category: {
      id: number;
      name: string;
    };
    subtypes: {
      id: number;
      name: string;
    }[];
  };
  admission_method: {
    method_description: string;
    subject_ratio: number | null;
    document_ratio: number | null;
    interview_ratio: number | null;
    practical_ratio: number | null;
    other_details: string | null;
    second_stage_first_ratio: number | null;
    second_stage_interview_ratio: number | null;
    second_stage_other_ratio: number | null;
    second_stage_other_details: string | null;
    eligibility: string;
    school_record_evaluation_score: string | null;
    school_record_evaluation_elements: string | null;
  };
  minimum_grade: {
    is_applied: "Y" | "N";
    description: string | null;
  } | null;
  interview: {
    is_reflected: number;
    interview_type: string | null;
    materials_used: string | null;
    interview_process: string | null;
    evaluation_content: string | null;
    interview_date: string | null;
    interview_time: string | null;
  } | null;
  previous_results: Array<{
    year: number;
    result_criteria: string;
    grade_cut: number | null;
    converted_score_cut: number | null;
    competition_ratio: number | null;
    recruitment_number: number | null;
  }>;
}